Items: Move ItemCache into dedicated submodule
