+++ cobbler/items/abstract/__init__.py
@@ -1,0 +1,3 @@
+"""
+Package to describe the abstract items that we define.
+"""
